broke out and refactored tests to fit new structure
